DBMS Quiz 추가

+++ src/InterfaceQuiz/Beauty.java
@@ -1,12 +1,14 @@
 package InterfaceQuiz;
 
+import java.math.BigDecimal;
+
 public class Beauty extends Product {
-    public Beauty(String name, int price, int weight) {
+    public Beauty(String name, BigDecimal price, int weight) {
         super(name, price, weight);
     }
 
     @Override
-    public int getDiscountAmount() {
-        return 10000;
+    public BigDecimal getDiscountAmount() {
+        return BigDecimal.valueOf(10000);
     }
 }
