+++ src/InterfaceQuiz/LargeAppliance.java
@@ -1,7 +1,9 @@
 package InterfaceQuiz;
 
+import java.math.BigDecimal;
+
 public class LargeAppliance extends Product {
-    public LargeAppliance(String name, int price, int weight) {
+    public LargeAppliance(String name, BigDecimal price, int weight) {
         super(name, price, weight);
     }
 }
