+++ src/InterfaceQuiz/Main.java
@@ -1,10 +1,12 @@
 package InterfaceQuiz;
 
+import java.math.BigDecimal;
+
 public class Main {
     public static void main(String[] args) {
-        Product beauty = new Beauty("beauty", 30000, 2);
-        Product grocery = new Grocery("grocery", 20000, 3);
-        Product largeAppliance = new LargeAppliance("largeAppliance", 50000, 5);
+        Product beauty = new Beauty("beauty", BigDecimal.valueOf(30000), 2);
+        Product grocery = new Grocery("grocery", BigDecimal.valueOf(20000), 3);
+        Product largeAppliance = new LargeAppliance("largeAppliance", BigDecimal.valueOf(50000), 5);
 
         Cart cart = new Cart(new Product[]{beauty, grocery, largeAppliance});
         int totalDeliveryCharge = cart.calculateDeliveryCharge();
