+++ src/InterfaceQuiz/Promotion.java
@@ -1,6 +1,8 @@
 package InterfaceQuiz;
 
+import java.math.BigDecimal;
+
 public interface Promotion{
     // 할인 금액을 리턴하는 메소드
-    int getDiscountAmount();
+    BigDecimal getDiscountAmount();
 }
